for #7 now lets go eyes free
Want to open your garage door on voice command, or turn on the coffee machine from your bed?

Lets use the Voice Recognition intent to have the andoid understand you... or at least tell you what it thought you said. This commit shows adding a couple of functions to our make it talk code, to make it listen and repeat.